+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), even though this is not a library, so there will eventually be breaking changes in minor version updates, managed by the team.
 
+## [0.3.2] - 2021-11-03
+- Create a DB for each environment
+## [0.3.1] - 2021-11-03
+### Added
+- Finished base of CI/CD implementation
 ## [0.3.0] - 2021-08-12
 ### Added
 - Create a router and service to reset password